is_simple_vector -> utils
